Moved PR additions to PermissionsHandler.kt to new file CitraDirectoryUtils.kt
