+++ src/android/app/src/main/java/org/citra/citra_emu/utils/PermissionsHandler.kt
@@ -1,4 +1,4 @@
-// Copyright Citra Emulator Project / Azahar Emulator Project
+// Copyright 2023 Citra Emulator Project
 // Licensed under GPLv2 or any later version
 // Refer to the license.txt file included.
 
@@ -14,39 +14,8 @@ import org.citra.citra_emu.CitraApplication
 
 object PermissionsHandler {
     const val CITRA_DIRECTORY = "CITRA_DIRECTORY"
-    const val LIME3DS_DIRECTORY = "LIME3DS_DIRECTORY"
     val preferences: SharedPreferences =
         PreferenceManager.getDefaultSharedPreferences(CitraApplication.appContext)
-
-    fun needToUpdateManually(): Boolean {
-        val directoryString = preferences.getString(CITRA_DIRECTORY, "")
-        val limeDirectoryString = preferences.getString(LIME3DS_DIRECTORY,"")
-        return (directoryString != "" && limeDirectoryString != "" && directoryString != limeDirectoryString)
-    }
-
-    fun attemptAutomaticUpdateDirectory() {
-        val directoryString = preferences.getString(CITRA_DIRECTORY, "")
-        val limeDirectoryString = preferences.getString(LIME3DS_DIRECTORY,"")
-        if (needToUpdateManually()) {
-            return;
-        }
-       if (directoryString == "" && limeDirectoryString != "") {
-            // Upgrade from Lime3DS to Azahar
-            setCitraDirectory(limeDirectoryString)
-            removeLimeDirectoryPreference()
-            DirectoryInitialization.resetCitraDirectoryState()
-            DirectoryInitialization.start()
-
-       } else if (directoryString != "" && directoryString == limeDirectoryString) {
-            // Both the Lime3DS and Azahar directories are the same,
-            // so delete the obsolete Lime3DS value.
-            removeLimeDirectoryPreference()
-        }
-    }
-
-    fun removeLimeDirectoryPreference() {
-        preferences.edit().remove(LIME3DS_DIRECTORY).apply()
-    }
 
     fun hasWriteAccess(context: Context): Boolean {
         try {
